implementacao básica da api de configuracao de cursos

+++ ChurchSchool.App/App/CourseConfiguration.cs
@@ -1,10 +1,9 @@
 ﻿using ChurchSchool.Application.Contracts;
 using ChurchSchool.Domain.Contracts;
+using ChurchSchool.Domain.Entities;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace ChurchSchool.Application.App
 {
@@ -12,9 +11,12 @@ namespace ChurchSchool.Application.App
     {
         private ICourseConfigurationRepository _courseConfigurationRepository;
 
-        public CourseConfiguration(ICourseConfigurationRepository courseConfigurationRepository)
+        private ICourseRepository _courseRepository;
+
+        public CourseConfiguration(ICourseConfigurationRepository courseConfigurationRepository, ICourseRepository courseRepository)
         {
             _courseConfigurationRepository = courseConfigurationRepository;
+            _courseRepository = courseRepository;
         }
 
         public IEnumerable<Domain.Entities.CourseConfiguration> GetAll()
@@ -24,6 +26,22 @@ namespace ChurchSchool.Application.App
 
         public Domain.Entities.CourseConfiguration Add(Domain.Entities.CourseConfiguration entity)
         {
+            if (!entity.IsValid())
+                return entity;
+
+            if (entity.IsCurrentConfiguration)
+            {
+                try
+                {
+                    DisableCurrentConfiguration(entity);
+                }
+                catch (Exception ex)
+                {
+                    entity.AddError(ex.Message);
+                    return entity;
+                }
+            }
+
             return _courseConfigurationRepository.Add(entity);
         }
 
@@ -39,14 +57,80 @@ namespace ChurchSchool.Application.App
             return _courseConfigurationRepository.Filter(new Domain.Entities.CourseConfiguration { CourseId = courseId });
         }
 
-        public bool Remove(Guid id)
+        public ValidationResult Remove(Guid id)
         {
-            return _courseConfigurationRepository.Remove(id);
+            if (VerifyRelatedCourses(id))
+            {
+                var validation = new ValidationResult();
+                validation.AddError("Existem cursos ativos atrelados a esta configuração, mude a configuração nos cursos e tente novamente.");
+                return validation;
+            }
+
+            _courseConfigurationRepository.Remove(id);
+
+            return new ValidationResult();
         }
 
-        public bool Update(Domain.Entities.CourseConfiguration entity)
+        public ValidationResult Update(Domain.Entities.CourseConfiguration entity)
         {
-            return _courseConfigurationRepository.Update(entity);
+            if (entity == null)
+            {
+                var result = new ValidationResult();
+                result.AddError("Nenhuma configuraçao de curso informada para atualização.");
+                return result;
+            }
+
+            if (!entity.IsValid())
+                return entity;
+
+            if (entity.IsCurrentConfiguration)
+            {
+                try
+                {
+                    DisableCurrentConfiguration(entity);
+                }
+                catch (Exception ex)
+                {
+                    entity.AddError(ex.Message);
+                    return entity;
+                }
+            }
+
+            entity.UpdatedDate = DateTime.Now;
+
+            _courseConfigurationRepository.Update(entity);
+
+            return new ValidationResult();
+        }
+
+        private bool VerifyRelatedCourses(Guid id)
+        {
+            var result = _courseConfigurationRepository.Filter(new Domain.Entities.CourseConfiguration { Id = id });
+
+            return (result != null && result.Any(y => y.RelatedCourse != null));
+        }
+
+        private void DisableCurrentConfiguration(Domain.Entities.CourseConfiguration entity)
+        {
+            var course = _courseRepository.Filter(new Domain.Entities.Course { Id = entity.CourseId }).FirstOrDefault();
+
+            if (course != null)
+            {
+                if (course.RemovedDate != null)
+                {
+                    throw new Exception("Não é possível adicionar uma configuração a um curso desabilitado");
+                }
+
+                var configurations = _courseConfigurationRepository.GetByCourse(entity.CourseId);
+
+                foreach (var configuration in configurations)
+                {
+                    configuration.IsCurrentConfiguration = false;
+                }
+
+                course.CurrentConfiguration = entity;
+                _courseRepository.Update(course);
+            }
         }
     }
 }
